extract editorial resolver from n8n adapter
